+++ src/test/java/it/tredi/msa/test/EmlExtractionTest.java
@@ -2,10 +2,12 @@ package it.tredi.msa.test;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.util.List;
 
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -13,6 +15,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.util.ResourceUtils;
 
 import it.tredi.msa.mailboxmanager.ParsedMessage;
+import it.tredi.msa.mailboxmanager.docway.DocwayParsedMessage;
 import it.tredi.msa.test.conf.MsaTesterApplication;
 
 /**
@@ -126,6 +129,79 @@ public class EmlExtractionTest extends EmlReader {
 	}
 	
 	/**
+	 * Parsing di parti del messaggio incluse come SharedByteArrayInputStream
+	 * @throws Exception
+	 */
+	@Test
+	public void classCastExceptionExtraction() throws Exception {
+		String fileName = " java.lang.ClassCastException.eml";
+		File file = ResourceUtils.getFile("classpath:" + EML_LOCATION + "/" + fileName);
+		
+		System.out.println("input file = " + fileName);
+		
+		ParsedMessage parsed = new ParsedMessage(readEmlFile(file));
+		
+		assertNotNull(parsed);
+		assertNotNull(parsed.getMessageId());
+		
+		System.out.println("messageId = " + parsed.getMessageId());
+		System.out.println("subject = " + parsed.getSubject());
+		System.out.println("from address = " + parsed.getFromAddress());
+		
+		boolean filefound = false;
+		List<String> attachments = parsed.getAttachmentsName();
+		System.out.println("attachments count = " + attachments.size());
+		for (String name : attachments) {
+			System.out.println("\tattach name = " + name);
+			if (name.equals("Istituto Fiduciario Lombardo SpA.pdf"))
+				filefound = true;
+		}
+		
+		assertEquals(3, parsed.getAttachments().size());
+		assertTrue(filefound);
+		
+		String fromDatiCert = parsed.getMittenteAddressFromDatiCertPec();
+		System.out.println("from dati cert = " + fromDatiCert);
+		
+		assertNotNull(fromDatiCert);
+		assertEquals("legal@pec.lacolombofinanziaria.com", fromDatiCert);
+	}
+	
+	/**
+	 * Estrazione dati da messaggio contenente molteplici istanze del file daticert.xml (inoltri vari di email)
+	 * @throws Exception
+	 */
+	@Test
+	public void nullPointerMultiDatiCert2Extraction() throws Exception {
+		String fileName = "nullPointer_multi_daticert_2.eml";
+		File file = ResourceUtils.getFile("classpath:" + EML_LOCATION + "/" + fileName);
+		
+		System.out.println("input file = " + fileName);
+		
+		ParsedMessage parsed = new ParsedMessage(readEmlFile(file));
+		
+		assertNotNull(parsed);
+		assertNotNull(parsed.getMessageId());
+		
+		System.out.println("messageId = " + parsed.getMessageId());
+		System.out.println("subject = " + parsed.getSubject());
+		System.out.println("from address = " + parsed.getFromAddress());
+		
+		List<String> attachments = parsed.getAttachmentsName();
+		System.out.println("attachments count = " + attachments.size());
+		for (String name : attachments)
+			System.out.println("\tattach name = " + name);
+		
+		assertEquals(15, parsed.getAttachments().size());
+		
+		String fromDatiCert = parsed.getMittenteAddressFromDatiCertPec();
+		System.out.println("from dati cert = " + fromDatiCert);
+		
+		assertNotNull(fromDatiCert);
+		assertEquals("LOM.sospensioni.discarichi@pec.agenziariscossione.gov.it", fromDatiCert);
+	}
+	
+	/**
 	 * Estrazione dati da messaggio contenente spazi o caratteri di controllo su un indirizzo 
 	 * email: javax.mail.internet.AddressException: Domain contains control or whitespace
 	 * @throws Exception
@@ -199,6 +275,169 @@ public class EmlExtractionTest extends EmlReader {
 		
 		assertNotNull(fromDatiCert);
 		assertEquals("riccardoguerra@pec.ordineavvocatigrosseto.com", fromDatiCert);
+	}
+	
+	/**
+	 * Estrazione dati da messaggio in caso di errori di encoding all'interno del file daticert.xml
+	 * @throws Exception
+	 */
+	@Test
+	public void daticertEncodingExceptionExtraction() throws Exception {
+		/*
+		String fileName = null;
+		File file = null;
+		ParsedMessage parsed = null;
+		List<String> attachments = null;
+		String fromDatiCert = null;
+		*/
+		
+		// CASO 1
+		String fileName = "daticertEncoding.eml";
+		File file = ResourceUtils.getFile("classpath:" + EML_LOCATION + "/" + fileName);
+		
+		System.out.println("input file = " + fileName);
+		
+		ParsedMessage parsed = new ParsedMessage(readEmlFile(file, false));
+		
+		assertNotNull(parsed);
+		assertNotNull(parsed.getMessageId());
+		
+		System.out.println("messageId = " + parsed.getMessageId());
+		System.out.println("subject = " + parsed.getSubject());
+		System.out.println("from address = " + parsed.getFromAddress());
+		
+		List<String> attachments = parsed.getAttachmentsName();
+		System.out.println("attachments count = " + attachments.size());
+		for (String name : attachments)
+			System.out.println("\tattach name = " + name);
+		
+		assertEquals(3, parsed.getAttachments().size());
+		
+		System.out.println("to addresses = " + parsed.getToAddressesAsString());
+		assertNotNull(parsed.getToAddressesAsString());
+		
+		String fromDatiCert = parsed.getMittenteAddressFromDatiCertPec();
+		System.out.println("from dati cert = " + fromDatiCert);
+		
+		assertNotNull(fromDatiCert);
+		assertEquals("consulenza.legale@pec.bppb.it", fromDatiCert);
+		
+		// CASO 2
+		fileName = "daticertEncoding2.eml";
+		file = ResourceUtils.getFile("classpath:" + EML_LOCATION + "/" + fileName);
+		
+		System.out.println("input file = " + fileName);
+		
+		parsed = new ParsedMessage(readEmlFile(file, false));
+		
+		assertNotNull(parsed);
+		assertNotNull(parsed.getMessageId());
+		
+		System.out.println("messageId = " + parsed.getMessageId());
+		System.out.println("subject = " + parsed.getSubject());
+		System.out.println("from address = " + parsed.getFromAddress());
+		
+		attachments = parsed.getAttachmentsName();
+		System.out.println("attachments count = " + attachments.size());
+		for (String name : attachments)
+			System.out.println("\tattach name = " + name);
+		
+		assertEquals(3, parsed.getAttachments().size());
+		
+		System.out.println("to addresses = " + parsed.getToAddressesAsString());
+		assertNotNull(parsed.getToAddressesAsString());
+		
+		fromDatiCert = parsed.getMittenteAddressFromDatiCertPec();
+		System.out.println("from dati cert = " + fromDatiCert);
+		
+		assertNotNull(fromDatiCert);
+		assertEquals("pec@pec.easybc.it", fromDatiCert);
+	}
+	
+	/**
+	 * Estrazione dati da messaggio in caso di eccezione di encoding sul 
+	 * contenuto: java.io.UnsupportedEncodingException: utf-7
+	 * @throws Exception
+	 */
+	@Test
+	@Ignore // Riabilitare solo dopo la configurazione indicata a questo url: https://www.bottaioli.it/internet/java-io-unsupportedencodingexception-unicode-1-1-utf-7-solved/
+	public void utf7ErrorExtraction() throws Exception {
+		String fileName = "utf7.eml";
+		File file = ResourceUtils.getFile("classpath:" + EML_LOCATION + "/" + fileName);
+		
+		System.out.println("input file = " + fileName);
+		
+		ParsedMessage parsed = new ParsedMessage(readEmlFile(file, false));
+		
+		assertNotNull(parsed);
+		assertNotNull(parsed.getMessageId());
+		
+		System.out.println("messageId = " + parsed.getMessageId());
+		System.out.println("subject = " + parsed.getSubject());
+		System.out.println("from address = " + parsed.getFromAddress());
+		
+		List<String> attachments = parsed.getAttachmentsName();
+		System.out.println("attachments count = " + attachments.size());
+		for (String name : attachments)
+			System.out.println("\tattach name = " + name);
+		
+		assertEquals(3, parsed.getAttachments().size());
+		
+		System.out.println("to addresses = " + parsed.getToAddressesAsString());
+		assertNotNull(parsed.getToAddressesAsString());
+		
+		String fromDatiCert = parsed.getMittenteAddressFromDatiCertPec();
+		System.out.println("from dati cert = " + fromDatiCert);
+		
+		assertNotNull(fromDatiCert);
+		assertEquals("studiodurand@legalmail.it", fromDatiCert);
+		
+		String html = parsed.getHtmlParts();
+		System.out.println(html);
+		assertNotNull(html);
+		assertTrue(html.trim().startsWith("<HTML><HEAD><TITLE>POSTA CERTIFICATA: definizione TE4I SRL  08534231009</TITLE></HEAD>"));
+	}
+	
+	/**
+	 * Errore durante il parsing del file Eccezione.xml allegato al messaggio
+	 * @throws Exception
+	 */
+	@Test
+	public void eccezioneXmlParsingExceptionExtraction() throws Exception {
+		String fileName = "erroreParsingEccezioneXML.eml";
+		File file = ResourceUtils.getFile("classpath:" + EML_LOCATION + "/" + fileName);
+		
+		System.out.println("input file = " + fileName);
+		
+		DocwayParsedMessage parsed = new DocwayParsedMessage(readEmlFile(file, false));
+		
+		assertNotNull(parsed);
+		assertNotNull(parsed.getMessageId());
+		
+		System.out.println("messageId = " + parsed.getMessageId());
+		System.out.println("subject = " + parsed.getSubject());
+		System.out.println("from address = " + parsed.getFromAddress());
+		
+		System.out.println("PEC message = " + parsed.isPecMessage());
+		System.out.println("PEC receipt = " + parsed.isPecReceipt());
+		
+		List<String> attachments = parsed.getAttachmentsName();
+		System.out.println("attachments count = " + attachments.size());
+		for (String name : attachments)
+			System.out.println("\tattach name = " + name);
+		
+		assertEquals(3, parsed.getAttachments().size());
+		
+		System.out.println("to addresses = " + parsed.getToAddressesAsString());
+		assertNotNull(parsed.getToAddressesAsString());
+		
+		String fromDatiCert = parsed.getMittenteAddressFromDatiCertPec();
+		System.out.println("from dati cert = " + fromDatiCert);
+		
+		assertNotNull(fromDatiCert);
+		assertEquals("protocollo@pec.comune.vignate.mi.it", fromDatiCert);
+		
+		assertTrue(parsed.isNotificaEccezioneInteropPAMessage("ADER", "ISC"));
 	}
 	
 }